requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
flake8==6.0.0
idna==3.4
importlib-metadata==6.1.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
zipp==3.15.0
